Added sveltekit backend (#7)
