+++ src/app.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8" />
+		<meta name="description" content="" />
+		<link rel="icon" href="%svelte.assets%/images/CAuth-logo.svg" />
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
+		<link rel="preload" as="font" href="https://fonts.gstatic.com/s/materialicons/v126/flUhRq6tzZclQEJ-Vdg-IuiaDsNc.woff2" type="font/woff2" crossorigin>
+		<link rel="preload" as="font" href="https://fonts.gstatic.com/s/oswald/v47/TK3iWkUHHAIjg752FD8Ghe4.woff2" type="font/woff2" crossorigin>
+		<link rel="preconnect" href="https://fonts.gstatic.com">
+    	<link rel="preload" as="style" onload="this.onload=null;this.rel='stylesheet'" href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,600;0,700;0,800;1,300;1,400;1,600;1,700;1,800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700&display=swap" crossorigin>
+		<link rel="preload" as="style" onload="this.onload=null;this.rel='stylesheet'" href="https://fonts.googleapis.com/css2?family=Oswald:wght@200;300;400;500;600;700&display=swap" crossorigin>
+		%svelte.head%
+	</head>
+	<body>
+		%svelte.body%
+	</body>
+</html>
