+++ src/routes/__error.svelte
@@ -1,0 +1,3 @@
+<div class="content">
+    <h1>Error 404</h1>
+</div>
