+++ src/routes/__layout.svelte
@@ -1,0 +1,9 @@
+<script>
+    import Header from "../components/Header.svelte"
+    import Content from "../components/Content.svelte"
+    import Footer from "../components/Footer.svelte"
+    import "../style/global.scss"
+</script>
+<Header />
+<Content><slot></slot></Content>
+<Footer />
