+++ src/routes/index.svelte
@@ -1,0 +1,53 @@
+<style lang="scss">
+    #home {
+        padding-top: 20px;
+        #goals {
+            display: flex;
+            justify-content: space-evenly;
+            .goal {
+                margin: 20px;
+                text-align: center;
+                width: 200px;
+            }
+        }
+    }
+    </style>
+    <div id="home" class="content">
+        <h1>Cryptographic Authentication - the next step for privacy and digital security in the Internet</h1>
+        <p>
+            Every single year there are news about hacked platforms, which are followed by the publication of private user data.
+            Unfortunately, it still happens very often that among the hacked platforms are some that store passwords unencrypted.
+            For the 21st century, the digital age, this is a mistake for which there are no excuses. <br>
+            <br>
+            Every time a user provides his or her personal data on a platform, he or she must trust that the platform will process and store the data responsibly.
+            Unfortunately, it becomes apparent every year that such responsible handling of sensitive data is not the norm.
+            Even tech giants like facebook have apparently stored passwords unencrypted for years, according to
+            <a href="https://krebsonsecurity.com/2019/03/facebook-stored-hundreds-of-millions-of-user-passwords-in-plain-text-for-years/">krebsonsecurity, 2019</a>. <br>
+            <br>
+            At C-Auth, we believe that security must always be a priority. That is why we firmly believe that such cases of published passwords should no longer exist.
+            Furthermore, it is not the platforms that should have control over personal data, but the users.
+            For this purpose, we are developing C-Auth. We have set ourselves goals that we want to achieve when using C-Auth:<br>
+        </p>
+        <div id="goals">
+            <div class="goal">
+                <img src="https://img.icons8.com/clouds/100/000000/security-lock-yellow.png" alt="security">
+                <h2>Security</h2>
+                <span>The data users provide should <b>always</b> be secure. No unauthorized instance should ever be able to obtain any of their data.</span>
+            </div>
+            <div class="goal">
+                <img src="https://img.icons8.com/clouds/100/000000/take-it-easy.png" alt="usability">
+                <h2>Usability</h2>
+                <span>Any tool related to C-Auth should be <b>easy to use</b> for anyone, also for people without technical knowledge.</span>
+            </div>
+            <div class="goal">
+                <img src="https://img.icons8.com/clouds/100/000000/joystick.png" alt="full control">
+                <h2>Full Control</h2>
+                <span>Users should have control over their data <b>at any time</b>. Not only which data they provide, but also who can access it.</span>
+            </div>
+            <div class="goal">
+                <img src="https://img.icons8.com/clouds/100/000000/file.png" alt="transparency">
+                <h2>Transparency</h2>
+                <span>Who accessed which data, and when? Users should be able to see that aswell</span>
+            </div>
+        </div>
+    </div>
