+++ src/stores/general.js
@@ -1,3 +1,3 @@
 import {writable} from 'svelte/store'
 
-export const loggedIn = writable(true);
+export const loggedIn = writable(false);
