--- vite.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from 'vite'
-import { svelte } from '@sveltejs/vite-plugin-svelte'
-import sveltePreProcess from 'svelte-preprocess'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [svelte({
-    preprocess: sveltePreProcess()
-  })]
-})
